+++ eox_nelp/one_time_password/api/v1/urls.py
@@ -1,0 +1,10 @@
+"""eox_nelp one_time_password api v1 urls"""
+from django.urls import path
+
+from eox_nelp.one_time_password.api.v1.views import generate_otp
+
+app_name = "eox_nelp"  # pylint: disable=invalid-name
+
+urlpatterns = [
+    path("generate/", generate_otp, name="generate-otp"),
+]
