+++ eox_nelp/urls.py
@@ -33,4 +33,5 @@ urlpatterns = [  # pylint: disable=invalid-name
     path('payment-notifications/', include('eox_nelp.payment_notifications.urls', namespace='payment-notifications')),
     path('api/user-profile/', include('eox_nelp.user_profile.api.urls', namespace='user-profile-api')),
     path('api/pearson_vue/', include('eox_nelp.pearson_vue.api.urls', namespace='pearson-vue-api')),
+    path('api/one-time-password/', include('eox_nelp.one_time_password.api.urls', namespace='one-time-password-api')),
 ]
