+++ eox_nelp/user_profile/api/v1/views.py
@@ -8,7 +8,6 @@ Function-views:
 """
 import logging
 
-from django.conf import settings
 from django.core.cache import cache
 from django.db import transaction
 from django.http import HttpResponseForbidden, JsonResponse
@@ -19,52 +18,8 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
 from eox_nelp.edxapp_wrapper.user_api import accounts, errors
-from eox_nelp.utils import generate_otp_code
 
 logger = logging.getLogger(__name__)
-
-
-@api_view(["POST"])
-@authentication_classes((
-    SessionAuthenticationAllowInactiveUser,
-))
-@permission_classes((IsAuthenticated,))
-def generate_otp(request):
-    """ View for generate OTP.
-    ## Usage
-
-    ### **POST** /eox-nelp/api/user-profile/v1/generate-otp/
-
-    request example data:
-    ``` json
-    {
-        "phone_number": 3213123123
-    }
-    ```
-    **POST Response Values**
-    ``` json
-    {
-        "message": "Success generate-otp!"
-    }
-    ```
-    """
-    user_phone_number = request.data.get("phone_number", None)
-
-    if not user_phone_number:
-        return JsonResponse(
-            data={"detail": "missing phone_number in data."},
-            status=status.HTTP_400_BAD_REQUEST,
-        )
-
-    otp = generate_otp_code(
-        length=getattr(settings, "PHONE_VALIDATION_OTP_LENGTH", 8),
-        custom_charset=getattr(settings, "PHONE_VALIDATION_OTP_CHARSET", ""),
-    )
-    user_otp_key = f"{request.user.username}-{user_phone_number}"
-    logger.info("generating otp %s*****", user_otp_key[:-5])
-    cache.set(user_otp_key, otp, timeout=getattr(settings, "PHONE_VALIDATION_OTP_TIMEOUT", 600))
-
-    return Response({"message": "Success generate-otp!"}, status=status.HTTP_201_CREATED)
 
 
 @api_view(["POST"])
