+++ eox_nelp/utils.py
@@ -1,7 +1,5 @@
 """Utils that can be used for the plugin project"""
-import random
 import re
-import string as string_module
 from copy import copy
 
 from opaque_keys.edx.keys import CourseKey
@@ -156,22 +154,3 @@ def camel_to_snake(string):
         String in snake case.
     """
     return re.sub(r'(?<!^)(?=[A-Z])', '_', string).lower()
-
-
-def generate_otp_code(length=8, custom_charset=""):
-    """Generates a random 8-digit (or custom length) alphanumeric OTP string.
-
-    Args:
-        length (int, optional): The desired length of the OTP. Defaults to 8.
-        custom_charset (string, optional): A string with chars that are used to generate the OTP code. If is a falsy
-        value the otp code would be generated with
-        ascii_letters + digits ('abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789').
-
-    Returns:
-        str: The generated OTP string.
-    """
-    allowed_chars = string_module.ascii_letters + string_module.digits
-    allowed_chars = custom_charset if custom_charset else allowed_chars
-    otp = ''.join(random.choice(allowed_chars) for _ in range(length))
-
-    return otp
